add modal to russian version
