+++ src/js/ru/modules/swiper.js
@@ -1,0 +1,24 @@
+export const swiper = new Swiper('.swiper', {
+  loop: false,
+  slidesPerView: 1,
+  slidesPerGroup: 1,
+  breakpoints: {
+    670: {
+      slidesPerView: 2,
+      slidesPerGroup: 2,
+    },
+    1000: {
+      slidesPerView: 3,
+      slidesPerGroup: 3,
+    },
+  },
+  pagination: {
+    el: '.swiper-pagination',
+    clickable: true,
+  },
+
+  navigation: {
+    nextEl: '.swiper-button-next',
+    prevEl: '.swiper-button-prev',
+  },
+});
