+++ src/js/ru/modules/typing.js
@@ -1,0 +1,41 @@
+export const typing = () => {
+  const typedTextSpan = document.querySelector('.typed-text');
+  const cursorSpan = document.querySelector('.cursor');
+
+  const textArray = ['Сергей Брыкалов'];
+  const typingDelay = 200;
+  const erasingDelay = 100;
+  const newTextDelay = 5000;
+  let textArrayIndex = 0;
+  let charIndex = 0;
+
+  function type() {
+    if (charIndex < textArray[textArrayIndex].length) {
+      if (!cursorSpan.classList.contains('typing')) cursorSpan.classList.add('typing');
+      typedTextSpan.textContent += textArray[textArrayIndex].charAt(charIndex);
+      charIndex++;
+      setTimeout(type, typingDelay);
+    } else {
+      cursorSpan.classList.remove('typing');
+      setTimeout(erase, newTextDelay);
+    }
+  }
+
+  function erase() {
+    if (charIndex > 0) {
+      if (!cursorSpan.classList.contains('typing')) cursorSpan.classList.add('typing');
+      typedTextSpan.textContent = textArray[textArrayIndex].substring(0, charIndex - 1);
+      charIndex--;
+      setTimeout(erase, erasingDelay);
+    } else {
+      cursorSpan.classList.remove('typing');
+      textArrayIndex++;
+      if (textArrayIndex >= textArray.length) textArrayIndex = 0;
+      setTimeout(type, typingDelay + 2700);
+    }
+  }
+
+  document.addEventListener('DOMContentLoaded', function () {
+    if (textArray.length) setTimeout(type, 300);
+  });
+};
