+++ src/ru.html
@@ -12,7 +12,7 @@
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css" />
   <link rel="stylesheet" href="./scss/style.scss">
   <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js" defer></script>
-  <script src="./js/ru.js" defer></script>
+  <script src="./js/ru/ru.js" defer></script>
   <title>Сергей Брыкалов - Веб-Разработчик</title>
 </head>
 
@@ -99,74 +99,52 @@
       <div class="swiper">
         <div class="swiper-wrapper">
           <div class="swiper-slide">
-            <a href="https://sebryk.github.io/quickdrawing-react-app/#/" class="swiper-slide-link" target="_blank"
-              rel="noopener">
+            <div>
               <picture>
                 <source srcset="./images/portfolio/quickdrawing.webp">
                 <source srcset="./images/portfolio/quickdrawing.jpg">
-                <img class="swiper-slide-img" src="./images/portfolio/quickdrawing.webp"
+                <img class="swiper-slide-img" id="quickdrawing" src="./images/portfolio/quickdrawing.webp"
                   alt="Preview Image of Quickdrawing React App">
               </picture>
-              <a href="https://github.com/sebryk/quickdrawing-react-app" class="swiper-slide-btn"
-                target="_blank">GitHub</a>
-            </a>
+            </div>
           </div>
           <div class="swiper-slide">
-            <a href="https://wavetake.com" class="swiper-slide-link" target="_blank" rel="noopener">
-              <picture>
-                <source srcset="./images/portfolio/wavetake.webp">
-                <source srcset="./images/portfolio/wavetake.jpg">
-                <img class="swiper-slide-img" src="./images/portfolio/wavetake.webp"
-                  alt="Preview Image of Wavetake Site">
-              </picture>
-              <a href="https://github.com/sebryk/wavetake" class="swiper-slide-btn" target="_blank">GitHub</a>
-            </a>
+            <picture>
+              <source srcset="./images/portfolio/wavetake.webp">
+              <source srcset="./images/portfolio/wavetake.jpg">
+              <img class="swiper-slide-img" id="wavetake" src="./images/portfolio/wavetake.webp"
+                alt="Preview Image of Wavetake Site">
+            </picture>
           </div>
           <div class="swiper-slide">
-            <a href="https://sebryk.github.io/memory-cards-react-app/" class="swiper-slide-link" target="_blank"
-              rel="noopener">
-              <picture>
-                <source srcset="./images/portfolio/neon-memory-cards.webp">
-                <source srcset="./images/portfolio/neon-memory-cards.jpg">
-                <img class="swiper-slide-img" src="./images/portfolio/neon-memory-cards.webp"
-                  alt="Preview Image of Neon Memory Cards App">
-              </picture>
-              <a href="https://github.com/sebryk/memory-cards-react-app" class="swiper-slide-btn"
-                target="_blank">GitHub</a>
-            </a>
+            <picture>
+              <source srcset="./images/portfolio/neon-memory-cards.webp">
+              <source srcset="./images/portfolio/neon-memory-cards.jpg">
+              <img class="swiper-slide-img" id="neon-memory-cards" src="./images/portfolio/neon-memory-cards.webp"
+                alt="Preview Image of Neon Memory Cards App">
+            </picture>
           </div>
           <div class="swiper-slide">
-            <a href="https://sebryk.github.io/weather-react-app/" class="swiper-slide-link" target="_blank"
-              rel="noopener">
-              <picture>
-                <source srcset="./images/portfolio/weather.webp">
-                <source srcset="./images/portfolio/weather.jpg">
-                <img class="swiper-slide-img" src="./images/portfolio/weather.webp"
-                  alt="Preview Image of Weather React App">
-              </picture>
-              <a href="https://github.com/sebryk/weather-react-app" class="swiper-slide-btn" target="_blank">GitHub</a>
-            </a>
+            <picture>
+              <source srcset="./images/portfolio/weather.webp">
+              <source srcset="./images/portfolio/weather.jpg">
+              <img class="swiper-slide-img" id="weather" src="./images/portfolio/weather.webp"
+                alt="Preview Image of Weather React App">
+            </picture>
           </div>
           <div class="swiper-slide">
-            <a href="https://sebryk.github.io/locus/" class="swiper-slide-link" target="_blank" rel="noopener">
-              <picture>
-                <source srcset="./images/portfolio/locus.webp">
-                <source srcset="./images/portfolio/locus.jpg">
-                <img class="swiper-slide-img" src="./images/portfolio/locus.webp" alt="Preview Image of Locus Site ">
-              </picture>
-              <a href="https://github.com/sebryk/locus" class="swiper-slide-btn" target="_blank">GitHub</a>
-            </a>
+            <source srcset="./images/portfolio/locus.webp">
+            <source srcset="./images/portfolio/locus.jpg">
+            <img class="swiper-slide-img" id="locus" src="./images/portfolio/locus.webp"
+              alt="Preview Image of Locus Site ">
+            </picture>
           </div>
           <div class="swiper-slide">
-            <a href="https://sebryk.github.io/drum_machine/" class="swiper-slide-link" target="_blank" rel="noopener">
-              <picture>
-                <source srcset="./images/portfolio/drum-sampler.webp">
-                <source srcset="./images/portfolio/drum-sampler.jpg">
-                <img class="swiper-slide-img" src="./images/portfolio/drum-sampler.webp"
-                  alt="Preview Image of Drum Sampler App">
-              </picture>
-              <a href="https://github.com/sebryk/drum_machine" class="swiper-slide-btn" target="_blank">GitHub</a>
-            </a>
+            <source srcset="./images/portfolio/drum-sampler.webp">
+            <source srcset="./images/portfolio/drum-sampler.jpg">
+            <img class="swiper-slide-img" id="drum-sampler" src="./images/portfolio/drum-sampler.webp"
+              alt="Preview Image of Drum Sampler App">
+            </picture>
           </div>
           ...
         </div>
@@ -217,6 +195,8 @@
         2024
       </p>
     </footer>
+  </div>
+  <div class="modal">
   </div>
 </body>
